+++ demo/demoproject/tests.py
@@ -9,6 +9,8 @@ except ImportError:
     from django.core.urlresolvers import reverse
 
 from demoproject._compat import decode
+from demoproject.models import Meter
+from chartjs.util import value_or_null, NULL
 
 
 class LineChartJSTestCase(TestCase):
@@ -76,3 +78,43 @@ class HighChartJSTestCase(TestCase):
             json.loads(decode(resp.content))
         except ValueError:
             self.fail("%r is not valid json" % self.resp.content)
+
+
+class DiscontinuousDataTestCase(TestCase):
+
+    def setUp(self):
+        self.start_date = "2019-05-26"
+        self.end_date = "2019-06-04"
+        # water meter readings
+        Meter.objects.create(date="2019-05-26", name="water", reading=10)
+        Meter.objects.create(date="2019-05-27", name="water", reading=12)
+        Meter.objects.create(date="2019-05-28", name="water", reading=13)
+        Meter.objects.create(date="2019-05-29", name="water", reading=15)
+        Meter.objects.create(date="2019-06-01", name="water", reading=16)
+        Meter.objects.create(date="2019-06-02", name="water", reading=18)
+        Meter.objects.create(date="2019-06-03", name="water", reading=20)
+        Meter.objects.create(date="2019-06-04", name="water", reading=21)
+        # gas meter readings
+        Meter.objects.create(date="2019-05-28", name="gas", reading=15)
+        Meter.objects.create(date="2019-05-29", name="gas", reading=13)
+        Meter.objects.create(date="2019-05-30", name="gas", reading=12)
+        Meter.objects.create(date="2019-05-31", name="gas", reading=14)
+        Meter.objects.create(date="2019-06-01", name="gas", reading=16)
+        Meter.objects.create(date="2019-06-02", name="gas", reading=17)
+
+    def test_generator_fills_end_values_with_null(self):
+        queryset = Meter.objects.filter(name="gas")
+        actual_data = []
+        for item in value_or_null(self.start_date, self.end_date, queryset, "date", "reading"):
+            actual_data.append(item)
+        expected_data = [NULL, NULL, 15, 13, 12, 14, 16, 17, NULL, NULL]
+        self.assertEqual(actual_data, expected_data)
+
+    def test_generator_fills_middle_values_with_null(self):
+        queryset = Meter.objects.filter(name="water")
+        actual_data = []
+        for item in value_or_null(self.start_date, self.end_date, queryset, "date", "reading"):
+            actual_data.append(item)
+        expected_data = [10, 12, 13, 15, NULL, NULL, 16, 18, 20, 21]
+        self.assertEqual(actual_data, expected_data)
+