+++ demo/demoproject/views.py
@@ -8,6 +8,9 @@ from chartjs.colors import next_color, COLORS
 from chartjs.views.columns import BaseColumnsHighChartsView
 from chartjs.views.lines import BaseLineChartView, HighchartPlotLineChartView
 from chartjs.views.pie import HighChartPieView, HighChartDonutView
+from chartjs.util import date_range, value_or_null
+
+from demoproject.models import Meter
 
 
 class ColorsView(TemplateView):
@@ -78,6 +81,49 @@ class DonutHighChartJSONView(ChartMixin, HighChartDonutView):
     pass
 
 
+class DiscontinuousDatesChartJSONView(ChartMixin, BaseLineChartView):
+    start_date = "2019-05-26"
+    end_date = "2019-06-04"
+
+    def get_providers(self):
+        return ["Water", "Gas"]
+
+    def get_labels(self):
+        return [dt for dt in date_range(self.start_date, self.end_date)]
+
+    def get_data(self):
+        result = []
+        water = Meter.objects.filter(name="water")
+        data = [item for item in value_or_null(self.start_date, self.end_date, water, "date", "reading")]
+        result.append(data)
+        gas = Meter.objects.filter(name="gas")
+        data = [item for item in value_or_null(self.start_date, self.end_date, gas, "date", "reading")]
+        result.append(data)
+        return result
+
+class DiscontinuousDatesHighChartJSONView(ChartMixin, HighchartPlotLineChartView):
+    title = _('Discontinuous Line HighChart Test')
+    y_axis_title = _('Volume')
+    start_date = "2019-05-26"
+    end_date = "2019-06-04"
+
+    def get_providers(self):
+        return ["Water", "Gas"]
+
+    def get_labels(self):
+        return [dt for dt in date_range(self.start_date, self.end_date)]
+
+    def get_data(self):
+        result = []
+        water = Meter.objects.filter(name="water")
+        data = [item for item in value_or_null(self.start_date, self.end_date, water, "date", "reading")]
+        result.append(data)
+        gas = Meter.objects.filter(name="gas")
+        data = [item for item in value_or_null(self.start_date, self.end_date, gas, "date", "reading")]
+        result.append(data)
+        return result
+
+
 # Pre-configured views.
 colors = ColorsView.as_view()
 
@@ -87,3 +133,5 @@ line_chart_json = LineChartJSONView.as_view()
 line_highchart_json = LineHighChartJSONView.as_view()
 pie_highchart_json = PieHighChartJSONView.as_view()
 donut_highchart_json = DonutHighChartJSONView.as_view()
+discontinuous_dates_chart_json = DiscontinuousDatesChartJSONView.as_view()
+discontinuous_dates_highchart_json = DiscontinuousDatesHighChartJSONView.as_view()